# api/openapi.yaml
openapi: 3.0.3
...
info:
  title: microvm
  version: "1.0.0"
  description: |
    Every sandbox is a real virtual machine with its own kernel. Code inside one
    is assumed hostile.

    There are two ways to run code, and the difference is the point:

    - **Sandboxes** are held. You create one, run commands in it, and it is
      yours until you stop it or its TTL elapses. Creation fails when the node
      is full, so backpressure is yours to handle.
    - **Tasks** are queued. You hand work to the fleet and it runs when a slot
      frees anywhere. Submission never fails for capacity.

    Use a sandbox for several commands sharing state; use a task for throughput.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080/v1
    description: A local daemon. Bound to loopback by default, because this API creates VMs that run arbitrary code.
security:
  - bearerAuth: []
...
paths:
...
  /sandboxes/{sandbox}/files:
    parameters:
      - $ref: '#/components/parameters/SandboxId'
    post:
      operationId: createFile
      summary: Upload a file
      description: Writes a file into the sandbox, creating parent directories.
      tags: [Files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FileCreateParams' }
      responses:
        '201':
          description: The file was written.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/File' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }

    get:
      operationId: retrieveFile
      summary: Download a file
      description: |
        The path is a query parameter rather than part of the route because a
        file path contains slashes, and encoding them into a path segment would
        make every caller escape their own paths to name a file.
      tags: [Files]
      parameters:
        - name: path
          in: query
          required: true
          description: Absolute path inside the sandbox.
          schema: { type: string, example: "/app/out.json" }
      responses:
        '200':
          description: The file's contents.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/CapacityExhausted' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>`. Tokens are compared in constant time.
  parameters:
    SandboxId:
      name: sandbox
      in: path
      required: true
      description: |
        A sandbox belongs to the tenant whose token created it, and every route
        below resolves it against the caller: another tenant's sandbox is a `404`,
        not a `403`. An ID is not a capability, and a `403` would confirm which of a
        guessed range exist. So a `404` here means one of three things -- never
        existed, already forgotten, or not yours -- and the API will not say which.
      schema: { type: string, pattern: '^sb_[0-9A-HJKMNP-TV-Z]{26}$', example: "sb_01JZ8QK3M4N5P6R7S8T9V0W1X2" }
...
  responses:
    InvalidRequest:
      description: The request is malformed or contradicts itself.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: The token is missing or wrong.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFound:
      description: No such object.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: The object already exists, or is in a state that forbids this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    CapacityExhausted:
      description: |
        No room right now. Either the node has no free slot, or the caller is over
        a per-tenant limit -- both are `capacity_error` with code
        `node_at_capacity`, deliberately, because the caller's move is the same and
        that is the pair every SDK already backs off on. The message is what tells
        them apart: a full node frees up on its own, whereas being over your own
        sandbox cap needs you to delete one.
      headers:
        Retry-After:
          description: |
            Seconds to wait, on the refusals whose end can be named to the second --
            the request-rate limit. Absent for a full node and for the per-tenant
            sandbox cap: when a slot frees is not something this node knows, and a
            number invented here would be a guess dressed as a fact.
          required: false
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      description: The body of every non-2xx reply.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/Error' }
    Error:
      type: object
      required: [type, code, message]
      properties:
        type:
          $ref: '#/components/schemas/ErrorType'
        code:
          type: string
          description: |
            A stable, machine-readable identifier. Switch on this. The message is
            for humans and may be reworded at any time.
          example: "sandbox_not_found"
        message:
          type: string
          description: A human-readable explanation. Never parse it.
          example: "No such sandbox: sb_01JZ8QK3M4N5P6R7S8T9V0W1X2"
        param:
          type: string
          description: Which request field was at fault, when one was.
          example: "image"
        request_id:
          type: string
          description: Quote this when reporting a problem.
    ErrorType:
      type: string
      description: |
        The class of error, which determines what a client should *do*:

        - `invalid_request_error` — the request is wrong. Retrying it unchanged
          will fail identically.
        - `authentication_error` — the token is missing or wrong.
        - `capacity_error` — no room right now. Retrying later may work; a task
          would not have failed at all.
        - `idempotency_error` — an Idempotency-Key was reused with a different
          body.
        - `api_error` — our fault. Safe to retry with the same Idempotency-Key.
      enum:
        - invalid_request_error
        - authentication_error
        - capacity_error
        - idempotency_error
        - api_error

    # --- pagination ---------------------------------------------------------
...
    File:
      type: object
      required: [object, path, size_bytes]
      properties:
        object: { type: string, enum: [file] }
        path: { type: string, example: "/app/main.py" }
        size_bytes: { type: integer }
        mode:
          type: string
          description: |
            The permission bits the file ended up with, octal. Set explicitly
            rather than left to whatever umask the guest happens to have, so this
            is what was asked for and not an approximation of it.
          example: "0644"
    FileCreateParams:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
          description: Absolute path inside the sandbox. Parent directories are created.
          example: "/app/main.py"
        content:
          type: string
          format: byte
          description: The file's bytes, base64 encoded so binary survives the trip.
        mode:
          type: string
          pattern: '^0?[0-7]{3}$'
          default: "0644"
          description: |
            Permission bits, octal -- `0755` for something you intend to execute.
            Defaults to `0644`, so an upload is readable and not runnable unless
            you say otherwise.

            Three digits, with an optional leading zero. setuid, setgid and the
            sticky bit are refused with an `invalid_request_error` on `mode`: the
            host writes this file as root, and the one thing an upload must never
            be able to do is leave a root-owned setuid binary behind for whatever
            runs in the guest next. Code inside the sandbox can still chmod its
            own files -- then it is the guest's own doing, and stays inside the
            guest.
          example: "0755"
...